Created helper function folder for creating history and removed focus component function invocation upon switching class/state
